src/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aioify==0.3.2
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
bs4==0.0.2
cachetools==5.5.2
celery
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
exceptiongroup==1.3.0
fastapi==0.116.1
filetype==1.2.0
frozenlist==1.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hvac==2.3.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langchain-core==0.3.72
langchain-google-genai==2.1.8
langsmith==0.4.8
module-wrapper==0.2.4
multidict==6.6.3
numpy==2.2.6
oauthlib==3.3.1
orjson==3.11.0
packaging==25.0
pgvector==0.4.1
poetry-version==0.1.5
prompt-toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyyaml==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
starlette==0.47.1
stdlib-list==0.11.1
telethon==1.40.0
tenacity==9.1.2
tomlkit==0.5.11
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
zstandard==0.23.0
gevent
eventlet
pybloom-live
elasticsearch
more-itertools
sentence-transformers
